+++ Source/EventFlow/EventStores/EventStoreBase.cs
@@ -35,8 +35,8 @@ namespace EventFlow.EventStores
     {
         protected class AllCommittedEventsPage
         {
-            public GlobalPosition NextGlobalPosition { get; private set; }
-            public IReadOnlyCollection<ICommittedDomainEvent> CommittedDomainEvents { get; private set; }
+            public GlobalPosition NextGlobalPosition { get; }
+            public IReadOnlyCollection<ICommittedDomainEvent> CommittedDomainEvents { get; }
 
             public AllCommittedEventsPage(
                 GlobalPosition nextGlobalPosition,
@@ -47,11 +47,11 @@ namespace EventFlow.EventStores
             }
         }
 
-        protected ILog Log { get; private set; }
-        protected IAggregateFactory AggregateFactory { get; private set; }
-        protected IEventUpgradeManager EventUpgradeManager { get; private set; }
-        protected IEventJsonSerializer EventJsonSerializer { get; private set; }
-        protected IReadOnlyCollection<IMetadataProvider> MetadataProviders { get; private set; }
+        protected ILog Log { get; }
+        protected IAggregateFactory AggregateFactory { get; }
+        protected IEventUpgradeManager EventUpgradeManager { get; }
+        protected IEventJsonSerializer EventJsonSerializer { get; }
+        protected IReadOnlyCollection<IMetadataProvider> MetadataProviders { get; }
 
         protected EventStoreBase(
             ILog log,
@@ -70,6 +70,7 @@ namespace EventFlow.EventStores
         public virtual async Task<IReadOnlyCollection<IDomainEvent<TAggregate, TIdentity>>> StoreAsync<TAggregate, TIdentity>(
             TIdentity id,
             IReadOnlyCollection<IUncommittedEvent> uncommittedDomainEvents,
+            IMetadata metadata,
             CancellationToken cancellationToken)
             where TAggregate : IAggregateRoot<TIdentity>
             where TIdentity : IIdentity
@@ -87,19 +88,20 @@ namespace EventFlow.EventStores
                 id);
 
             var batchId = Guid.NewGuid().ToString();
-            var storeMetadata = new[]
-                {
-                    new KeyValuePair<string, string>(MetadataKeys.BatchId, batchId), 
-                };
+            var storeMetadata = metadata
+                .Concat(new []
+                    {
+                        new KeyValuePair<string, string>(MetadataKeys.BatchId, batchId), 
+                    });
 
             var serializedEvents = uncommittedDomainEvents
                 .Select(e =>
                     {
-                        var metadata = MetadataProviders
+                        var md = MetadataProviders
                             .SelectMany(p => p.ProvideMetadata<TAggregate, TIdentity>(id, e.AggregateEvent, e.Metadata))
                             .Concat(e.Metadata)
                             .Concat(storeMetadata);
-                        return EventJsonSerializer.Serialize(e.AggregateEvent, metadata);
+                        return EventJsonSerializer.Serialize(e.AggregateEvent, md);
                     })
                 .ToList();