+++ RELEASE_NOTES.md
@@ -2,6 +2,11 @@
 
  * Breaking: `EventFlowOptions AddDefaults(...)` now also adds event
    definitions
+ * Breaking: `ICommand.Id` is renamed to `ICommand.AggregateId` to make "room"
+   for the new `ICommand.CommandId` property. If commands are serialized, then
+   it _might_ be important verify that the serialization still works. EventFlow
+   _does not_ serialize commands, so no mitigation is provided. If the
+   `Command<,>` is used, make sure to use the correct protected constructor
  * New: [RabbitMQ](http://www.rabbitmq.com/) is now supported through the new
    NuGet package called `EventFlow.RabbitMQ` which enables domain events to be
    published to the bus