+++ RELEASE_NOTES.md
@@ -9,6 +9,15 @@
    and register a service that implements `ISubscribeSynchronousToAll`. Services
    that implement this will automatically be added using the
    `AddSubscribers(...)` or `AddDefaults(...)` extension to `EventFlowOptions`
+ * New: `AutofacAggregateFactory` added to enable using `Autofac` as aggregate
+   factory making injecting services into an aggregate root constructor 
+   possible. Use `UseAutofacAggregateFactory(...)` to configure and replace 
+   the default `Activator` based aggregate factory
+ * New: `EventFlowOptions AddAggregateRoots(...)` can be used to register 
+   `IAggregateRoot<>` components thereby eliminating the need to use 
+   `RegisterServices(...)`
+ * New: Added `IServiceRegistration RegisterType(...)` to simplify registering 
+   services by type.   
 
 ### New in 0.10.642 (released 2015-08-17)